finalize: refactor port logic and init target cluster
Introduce a new object called InitializeConfig. It is used to store
configuration for creating segments on the target cluster. We create a master,
standby and primaries based on the contents of this struct, and it is also
saved to the disk in config.json.

This refactor is done in preparation for finalize: upgrade mirrors