UpdatePostgresqlConf: harden implementation
The current implementation won't be idempotent if, say, we replace the
port number with a number that is a superstring of the previous one.

Example, replacing 5000 with 50000:

   port=5000 -> port=50000

The second run will match "port=5000" again:

   port=50000 -> port=500000

While fixing this, we discovered some additional anchoring bugs related
to the GUC name and comments, so we added a suite to fix them all. The
current test subsumes the previous test, which was simply checking
argument equality, so we've replaced it entirely.

Co-authored-by: Kalen Krempely <kkrempely@pivotal.io>