CI: introduce AllVersions concept
The template Data.AllVersions consists of a combined and deduplicated
list of source and target versions. Its utility in this patch is limited
to the declaration of binary resources in the pipeline, but this will
allow us to make the 6-to-6 job a mundane case instead of a special one.