make check: switch to `go test` from Ginkgo
A Ginkgo workflow is difficult to get working in the land of CI and Go
modules due to several low-priority bugs/misfeatures in `go get` and `go
install`. We've been talking about removing it for a while now, though,
so this is a good time to do so.

For now, do not cache test results, while we fix the integration test
cache behavior. Developers who want the cache speedup can just use `go
test` directly.

Co-authored-by: David Krieger <dkrieger@pivotal.io>