CI: do not modify go.mod during build
We want any "required" changes to go.mod to fail a CI build. This should
be taken care of on developer machines.

As a concrete example, if a developer checks in a change to a file that
adds a new package dependency, but doesn't actually remember to check in
the pinned version in go.mod, the CI will now explicitly fail instead of
silently pulling in the latest and continuing.

Co-authored-by: David Krieger <dkrieger@pivotal.io>