CI: switch X-to-Y tests to ports 6020-6040
Some intermittent failures appear to be related to port collisions in
the ephemeral range, so switch to the same port range used by the
install-tests fix in ff9e6406f.

Eventually, we can probably reserve the default 50432-* range on the VMs
instead (unlike install-tests, which is using containers), but while
we're trying to debug and reason about these intermittent failures,
consistency seems more helpful.