Add centos6 jobs
Previously we were only testing upgrade on centos7.  Now, all upgrade jobs are
also ran on centos6, and to limit the workload, there are serial groups so that
a centos6 job must finish first before the respective centos7 job can start.

Another strategy to limit the strain on resources was setting the ccp reap time
to 3 hours instead of the default 8 hours.

The only job that currently fails is 6-to-6-centos-6 and has been removed for
now.