revert: add verbose option
Always show revert substeps. Stream the directories being deleted to the CLI
when the verbose option is set.

Currently we are only streaming the directories that are being deleted on the
hub, but once we introduce a way to stream output to the hub from the agents,
we'll want to print the directories being deleted on each agent as well.