Add archive directory safety precaution
Issue is that it is possible that a malicious party could forge a
ArchiveLogDirectoryRequest message to move an unexpected directory.
Solution is to obmit the old directory from the message and instead rely
on the agents' and hub's known old directory (i.e. log location), which
they know because they are already logging to that location.