revert after a successful execute in copy mode
Add support to revert a cluster after a succesfully running execute in
copy mode. This included stopping the target cluster and starting the
source cluster.

Since we currently only support reverting after a successful initialize
and execute in copy mode a test was removed which was stopping before
the target cluster was created. Since the target cluster was not created
this was causing references to hub.Config.Target to fail which are
needed to check if the target cluster is running. Revert currently does
not yet support when iniitalize or execute fails part way through.

Co-authored-by: David Krieger <dkrieger@pivotal.io>
Co-authored-by: David Kimura <dkimura@pivotal.io>