add hub-port flag to initialize
This adds the hub-port flag to initialize with the default port being
7527.

This removes the GPUPGRADE_HUB_PORT environment variable which is no
longer used.

When starting the hub the port specifed in the configuration file is
used rather than specifying --port. This simplifies the order of
operations in bootstrapping and starting the hub, and allowing other CLI
commands to connect to the hub.

When connecting to the hub the port is read from the hub configuration.
This is not ideal since it is mixing CLI and hub responsibilities.