GetArchiveDirectoryName: inject timestamp directly
It's not clear from the API or documentation that the output of
GetArchiveDirectoryName() varies by wall clock time:

    a := GetArchiveDirectoryName()
    b := GetArchiveDirectoryName()

    if a != b {
	panic("not the same directory!") // sometimes succeeds
    }

This differs from many of the other "directory name" construction
patterns, so there's a good chance it will trip somebody up.

Take the timestamp as part of the input parameters to make it obvious
that, if you need to share the directory name between multiple locations
in code, you should either cache the resulting directory name or the
input timestamp.