initialize supports using a config file
This adds support to initialize to accept a configuration file. The
default configuration to be shipped is found in the gpupgrade_config
file.

We error out if the file flag and command line flags other than verbose
are used together. Only one method or the other should be used.

Our configuration syntax and style is similar to upstream, but differs
in the following ways:
- an equal sign is required  between parameter names and values
- values are passed directly with no quoting needed.
- throws an error when a parameter is specified multiple times