filter: Refactor replacing of single quote in WITH Clause
This commit refactors the replacement of single quotes in the WITH
clause of the partition tables. The earlier approach required defining
matches for all the combinations, which is cumbersome when there are
several such patterns.
In the new approach, the main driver file should just include the
replacement function which must be applied on a line.  The replacement
function should performed the desired transformation on the line and
return it. The replacement function hides all the complexity for regex
match and any other transformation logic.