filter: Add Regex to delete a line with START WITH n
With GPDB6, in the DDL of CREATE SEQUENCE, START WITH <num> line is
present, however in case of GPDB5, this line is ommitted.  For now,
delete the line containing START WITH without checking the INITIAL VALUE
line. If we get cases in the dump which requires that, we will have to
do a look ahead and then delete if applicable.