ci: install rsync in no-install tests and clean up its yaml
The rsync tests, currently in the agent soon to be moved to utils,
were being skipped in no-install tests because
rsync was not being installed.  We add rsync to the image.

Also, since we are now using go modules, we can eliminate some
lines of code that are now unnecessary.

Co-authored-by: Kalen Krempely <kkrempely@pivotal.io>