revert: after execute, print out the source version and archive dir.
This can assist the user in finding the logs.

We have existing code for this: the revert() code on the hub places the
required data into a map, which is sent back to the cli over gRPC at the
end of revert().

The cli can then interpret this information and present it to the user.