deleteStateDirectory: don't check contents before deleting state directory
State directory on the agents does not contain any content until
pg_upgrade check is run in initialize, but we support revert both after
initialize and execute. If initialize failed before pg_upgrade check,
there will be no contents in the state directory, so there is currently
no marker file which validates the state directory.

Also, before this commit, the check used to verify the config.json and
status.json files on the agent before deleting the state directory, but
these files don't exist on agents.