test: revert.bats now correctly validates that standby comes up properly
We implement the correct logic to ensure that the standby is up correctly:
that is WAL log is in sync with the master via pg_stat_replication.

The implementation is detailed, as we might need to wait up to 10 minutes
for the standby to come into sync.  In addition, if the standby disconnects
from the master but is still running, it might reconnect.