upgrade: add the dbid of the standby to the master in UpgradeMaster
When executing a 5-6 pg_upgrade, the source master should know the dbid of its
standby.  This is in general the correct semantic.

Furthermore, depending on how we do revert after execute, the standby may
never sync its WAL with the master and exit abruptly:

1). rsync based revert (currently done for link-mode)
Setting the standby-dbid isn't strictly necessary but is still corret.

2). gprecoverseg based revert (currently done for copy-mode)
Setting the standby-dbib is needed to sync the standby to the master.

For context, see:
https://groups.google.com/a/greenplum.org/d/msg/gpdb-dev/RunORQtf0X4/reX311xIAwAJ