exectest: add NewCommandMock for mock-style tests
Writing more integration-level tests with exectest is currently
difficult. Clients must write low-level command verifiers, most of which
end up looking like an assertion that the arguments match some slice of
strings, and this obfuscates the test purpose. Additionally, if the unit
under test needs to call more than one external command, the only way to
control them independently is to create multiple injection points, which
starts to unravel quickly.

These sorts of tests are typically less interested in exact control over
the Command; rather, they want to perform setups like "all calls to
gpstart should fail" or "if pgrep returns no results, bash should be
executed with this set of arguments." This is the domain of mock
frameworks, and we already have one of those.

Add a new type of exectest pattern: NewCommandMock() returns both an
exec.Command implementation, and a gomock object to spy on and control
that implementation.