greenplum: switch to exectest.NewCommandMock()
Consolidate all external command calls to a single injection point, and
control it using the new mock framework. This declutters the start/stop
tests considerably, at the cost of new lines of test boilerplate due to
the need for mock cleanup. (This can go away once we switch to the Go
1.14 featureset, with the use of t.Cleanup().)

While porting the start/stop tests, I fixed a small bug in the

    stop cluster detects that cluster is already shutdown

test, which incorrectly injected an error condition rather than a "no
results" condition, as would occur with an already-shut-down cluster.
The rest of the test remains the same.