revert: introduce a idempotence checker
Setting the DebugIdempotence flag in the hub configuration will cause
the hub to fail every revert substep exactly once. The corresponding
end-to-end test simply loops in a call to revert, checking that any
errors are due to the idempotence framework, until the step succeeds.

This tests only a subset of idempotence -- namely, that each individual
substep (and any of its AlwaysRun predecessors) can be re-run after
completion -- but in practice this catches a number of common bugs.

Co-authored-by: Bhuvnesh Chaudhary <bchaudhary@vmware.com>