upgrade: allow any "fatal" output in --check error context
When parsing pg_upgrade output on an error, we currently look for
"fatal" at the end of the line to use as the starting marker for error
context collection. pg_upgrade may soon change to have timing
information printed at the end of each line, so allow any instance of
the "fatal" string to trigger the context collection.

This may catch more information than we need (say, if a pg_upgrade check
in the future contains the word "fatal"), but it will never catch too
little. We plan to revisit how error collection from pg_upgrade is done
in the near future, anyway, and this keeps the coupling low in case the
pg_upgrade output needs to continue changing.