revert: mark explicitly skipped steps
Revert has some substeps that must always be run, but depending on the
state of the system, they may not need to perform any work. We want to
show a SKIPPED status in the UI for these cases as well.

Introduce a sentinel error, step.Skip, that substep implementations may
return in order to tell the framework that no work is to be done. The UI
will report SKIPPED and the substep will be marked COMPLETE in the
status store.

Co-authored-by: Bhuvnesh Chaudhary <bchaudhary@vmware.com>