initialize: allow CLI-only substeps to be skipped
Two substep implementations in initialize are skipped if they've already
run, but currently they're marked COMPLETE every time. The previous
commit didn't cover these cases because the substep implementations
aren't running on the hub.

Have substep.Finish() accept the step.Skip sentinel as a special case,
similar to the previous commit, and report SKIPPED in the UI when it's
encountered. Finish() will clear the sentinel error before returning so
that it's not inappropriately reported.

Co-authored-by: Bhuvnesh Chaudhary <bchaudhary@vmware.com>