user commanders Response instance rather than pointer
Don't use a pointer for commanders.ExecuteResponse,
commanders.FinalizeResponse, and commanders.RevertResponse since they may be
null if connecting to the hub fails. This will result in a null pointer as the
response is used when step.Complete is executed.

This was causing "subcommands return an error if the hub is not started" in
test/initialize.bats to fail with an exit status of 2 instead of 1.