add steps.json
This tracks overall step status in order to validate if a step has been
run or is allowed to run. This uses the same data structure and file
store as substeps.json for convenience. A substep enum STEP_STATUS has
been added to track the overall step status. This is not to be used as a
normal substep, but rather internally to track the overall step status.

Co-authored-by: David Krieger <dkrieger@vmware.com>