add confirmation prompt
Fail fast by adding the confirmation prompt "after" validating the
current step.

When initialize is called with --automatic then no confirmation prompt
is presented and gpupgrade proceeds.

The --automatic flag is not accepted for any other command except
initialize.

Add automatic to other commands, but as a hidden flag. This is used for
end to end tests to not wait for user input.

Don't print "Error:" when the user cancels upgrade