bats: finalize.bats can specify the MASTER_DATA_DIRECTORY
This part of the test makes sure that, after finalize, the target
cluster is synchronized.  It was previously looking up the
MASTER_DATA_DIRECTORY of the target cluster using gpugprade config.
That will not work after this PR, as finalize will stop the hub.

Instead, use the MASTER_DATA_DIRECTORY, which is that of the source
cluster when this test is called.  That is the same as the target's
MASTER_DATA_DIRECTORY after finalize.