revert: update RevertHelp text to match actual substeps run
The list of substeps was outdated; this is what it does now,
if all substeps are run. This assumes link mode and a 5X
source cluster that requires recoverseg. This seems like
the right thing to do because, if those extra steps are
not run, they are "done" as no-ops implicitly. That is,
the revert will work.