Replace precision values with .XX
Refactor: Make the rules for greenplum 6x to its separate file

This commit refactors the main filter driver and separates it in
modules so that it's not tightly coupled with greenplum 6x. One can
add a new rule for another version of greenplum and it will pick the
specified rules.

Add filtering to replace precision values with .XX

Add rules for filtering 5x and 6x dump for precision values

Update the upgrade cluster script