build: verify and upload gpupgrade rpms to GCP
We intend to use rpms to distibute both the open source("oss")
and enterprise("enterprise") versions of gpupgrade.

Build the rpms and verify their contents.  Note:

- we upload a version "0.4.0-1.el7.x86_64" right now based on the rpm
of greenplum-upgrade-0.4.0-1.el7.x86_64.rpm since that is how Concourse
versions a GCP resource
- we need to improve this versioning as that upload will continually be
replaced; that is, each build right now is not unique, but only based on
the git tag.  We update this infrequently.
- the verification of the rpms just sanity checks the contents; we will
be passing the rpms into the actual tests in the CI in a later PR

Co-authored-by: Kalen Krempely <kkrempely@vmware.com>