create Target cluster based off Source
Target cluster now represents the final state of the upgrade process.
This is actually the same as the source cluster, but for readablity
throughout the code we use a different object called target.

When creating a cluster don't return a pointer as this  allows us to
make a deep copy of the object. This is useful in the next commit to
have config.Source and config.Target be the same, but not pointing to
the exact same object. This allows us to set the correct gphome for
source and target since they are different objects.