finalize: re-order steps
Update the mirrors and standby early on and update the catalog and
directories last. This allows for much greater clarity and understanding
internally between when to use the intermediate versus target cluster
data structures. This calrity also helps with debugging and potential
support scenarios. This also allows us to save the most potentially
dangerous step until last. Thus, if an error does occur the cluster has
fully been upgraded and just the catalog or directories need to be
renamed, which if needed could be done manually.