log the database destination and connection URI
For debugging and support log the database connection destination and
connection URI.

Standardize on using sql.Open(), and on closing the connection. There is
no real need to test that the connection settings we are passing are
what we want. That is just testing our mock inputs. Rather all real
tests should be done at a lower level in the connection URI specific
tests.