cluster: move GreenplumRunner to cluster
Greenplum runner had its how struct and interface which duplicated much
of the information already on the cluster object. Thus, move it to the
cluster to remove the need to pass all that information in.

Remove unneeded tests that simple test what we are mocking. That is, the
lower level tests already do that.