finalize_checks: wait between stopping master and primaries
Wait for FTS to update gp_segment_configuration between stopping the
coordinator and the primaries before doing the failover to the standby.
There appears to be a bug that casus the standby to not correctly have
the latest information even though the table is present, and data is on
the segments.