Remove greenplum version check
This commit removes the step in which we check if the
greenplum versions are matching between coordinator and
segments. The check was done via ssh which was prone to
failure as described in the following issue:
https://github.com/greenplum-db/gpupgrade/issues/582

We still use the ssh based checks for gpupgrade version.
We cannot use grpc since this check needs to happen before
agents are brought up and a grpc call needs them running.

Co-authored-by: Kalen Krempely <kkrempely@vmware.com>