add dynamic_library_path parameter
The dynamic_library_path parameter has been added to allow customers to
specify the location of their extensions installed outside of the
default location. This GUC is needed by pg_upgrade when performing a pg_dump
and pg_restore in order to correctly load the correct extension .so
file.

Only set this GUC when it is specified and not the default. This is to
keep things simple and there is no need to set this GUC on the customers
system if the default locations are used.

Co-authored-by: Orhan Kislal <okislal@vmware.com>