update add-semver to work with first tagged commit
The first commit after tagging add-semver.bash was failing with “parts[1]:
unbound variable” since git describe was returning 1.2.0 rather than
1.2.0-0-g9212e0e1. Update the script to work with both forms of git describe
output.

This also removes the hard coded “-1” packing build number. This would be
needed if we ever plan to release a new package without code changes. However,
this is a highly unlikely scenario and one where we could simply bump the
version. If needed, it can be added back later.